requirements.txt
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
redis==6.2.0
flask_login==0.6.3
python-dotenv==1.1.0
gunicorn==23.0.0
Flask-SocketIO==5.3.6
eventlet==0.33.3